feat(ci): WP-8 — gate Omotenashi browser-QA vs loja Nuxt + operador Django
Omotenashi é 1ª classe: o cutover headless aposentou as páginas Django de
cliente, e o gate antigo (que as navegava) saiu do CI. Este WP reconstrói o
gate para a topologia real — loja Nuxt (cliente) + Django (operador).

- Workflow dedicado .github/workflows/omotenashi-gate.yml (Postgres + Python +
  Node + Chrome → make omotenashi-browser-ci; artifact de screenshots/report/
  logs). runtime-gate.yml: comentário stale atualizado p/ apontar o novo gate.
- Orquestração única local==CI (scripts/run_omotenashi_browser_ci.sh): migrate +
  seed + build/serve da loja Nuxt (:3100, BFF→Django) + Django (:8001) com
  SHOPMAN_STOREFRONT_BASE_URL→loja + QA estrita + teardown dos dois servidores.
- Runner (run_omotenashi_browser_qa.mjs): cookie de sessão QA em TODAS as origens
  da matriz; sessão = superusuário (operador) + order-access grants
  (shopman_order_access_refs via order_ref na matriz) p/ os pedidos dos checks de
  pagamento/tracking — DRF não lê request.user, o acesso a pedido é por grant de
  sessão. resolveNavigable pula superfícies não-servidas com log explícito (nunca
  falso-verde); flag auth_gated trata o gate de login do checkout como esperado.
- Matriz (omotenashi_qa.py): POS repontado p/ pos_links.pos_url() (novo módulo +
  knob SHOPMAN_POS_BASE_URL), pulado até a revisão do PDV (fase C); novos campos
  auth_gated + order_ref.

Bugs reais que o gate surfaceou e foram corrigidos:
- shop/services/kds.py: qty de componente de bundle era int*Decimal → Decimal no
  JSONField do ticket (quebrava seed no SQLite; Postgres mascarava) → int().
- tracking/[ref].vue + pedido/[ref]/pagamento.vue não repassavam o cookie no SSR
  (≠ páginas de conta) → renderizavam o fallback "não encontrado" p/ um cliente
  que PODE ver o pedido; corrigido (headers de cookie no SSR).

Verificado: make omotenashi-browser-ci → 12 pass, 0 review, 2 skipped (POS); os 3
pedidos-cliente renderizam o estado REAL (PIX/expiração+recuperação/pronto).
make test 1844 passed; make lint verde; vitest 217 passed.

+++ shopman/shop/services/kds.py
@@ -254,7 +254,10 @@ def _build_routable_items(lines: list[dict]) -> list[dict]:
                 routable_items.append({
                     "sku": comp_sku,
                     "name": f"{comp_name} ({name})",
-                    "qty": qty * comp_qty,
+                    # comp_qty is a DecimalField; the KDS line qty is an integral
+                    # count everywhere else (and must be JSON-serializable for the
+                    # ticket items JSONField — plain json.dumps rejects Decimal).
+                    "qty": int(qty * comp_qty),
                     "notes": notes,
                     "parent_sku": ln["sku"],
                     "line_id": line_id,
